Don't inherit SafeHandle for C# prop-having class
Some classes who inherits SafeHandle have setter/getter by C# property
and their handle is not usable without checking property values.
So to maintain handle is meaningless and it can be converted just when
needed.

For CipherParameters class, it doesn't have setter/getter with C#
property so SafeHandle here is used properly and not removed.

Change-Id: Ib75edf5af74d4a0fbf44fbb7c16b836eee92a790
Signed-off-by: Kyungwook Tak <k.tak@samsung.com>