Make internal for abstract classes and static ctor
CipherParameters and AesCipherParameters are only internally inherited
and don't want to be included in API. By defining internal default
constructor, it can be inheritd only by assembly internally.

All Managers are only have static members and only used as static so no
constructor is needed (and should not be for integrity). Define internal
default ctors to prevent being called

Change-Id: I3013bf26b5b014129fd412abc37dac7335686f6a
Signed-off-by: Kyungwook Tak <k.tak@samsung.com>