Enable creating FormsApplication based IMEs
Since the FormsApplication class automatically creates
normal application window when created, we will override
PreCreate() method to acquire IME window from InputMethod
capi and assign it to the MainWindow member variable.
This patch needs to be submitted after the patch is applied
that changes the access modifier of FormsApplication's
MainWindow variable to "protected".

Change-Id: I5940757d3cb4608c3e8d6e69c509c05866698ee6