[Tizen] Create another native handle to make sure Registry.Unregister works fine.
This is much simpler solution with less code maintenance as derived classes do not have to call Registry.Unregister(this);.
Only BaseHandle class will be responsible for Registry.Register and Registry.UnRegister which is much simpler and symmetric solution.
The only drawback is that we have to create an extra native handle.

Change-Id: I2e3ab7b3311befb25141ae0816b7d2d77a2e3485