[Multimedia] Fixed a bug of MetadataExtractor.
The issue was an extractor object is collected by GC while the object is still in use by internal members.
It happens only with the release binary and seems the compiler optimization is involved.
By this patch, the internal operation that causes an fatal error, keeps the extractor object until it ends.

Change-Id: Ie44b3207bef44ecd2b47f4f8bce9b919c2a70a40
Signed-off-by: coderhyme <jhyo.kim@samsung.com>