[EflSharp] Update Circle and efl cs files
Circle and efl cs files are updated based on the latest efl.

Clickable and Selectable interfaces are updated.
All Popup classes are merged to Popup and AlertPopup.
SliderInterval class is no longer used.

Temporarily, List item's HintSizeMin should be set in app code for now.
The HintSizeMin code will be removed when the internal size calculation
logic is fixed.