[NUI] Fix reparenting of Child bug
If a child is added to a different layout the transition finished
animation can cause it to be removed from the new parent.

Cause : The removal animation is added to the stack, the add animation is added to
the stack.  Animations are run.  When animations finish the itemRemovalQueue is run, this
removes the child as the removal code added it to the itemRemovalQueue.

Fix : In the case of replacement, set a flag on the child so the removal code can excluded it
from the itemRemovalQueue.

Change-Id: Ib7da9e446c990258c8415e584bbb624dae4d6fa5