[NUI] Use padding & margin to meausre size in Layout
Previously, padding of parent and margin of child are not concerned when measure child size.

For now, When child use MatchParent,
the size of child will be ParentSize - ParentPadding - childMargin.

And when child use WrapContent,
the size of child will be GrandchildSize + childPadding.

Also Weight will be use only ParentSize - ParentPadding - childMargin size for measuring.