[NUI] remove input method dependency from View.ControlState (#1585)
`ConstrolState` can be changed by API or Input(`Tizen.NUI.Key` and
`Tizen.NUI.Touch`). it means `ConstrolState` is not always related to
Input method. some Controls need it to use state changed effect. however,
not all controls require that. if we need some property to change state,
we can add it to Style Extension classes.