[NUI] Change suggested minimun size in Layout.
When user set maximum size or minimum size instaed of size and specification,
it means that user want to use natural size of component and makes it bigger than minimum and smaller than maximum.

Currently, when using text, it lays whole text on one line and returns its size as natural size even if multiline is enabled.
So, we need to calculate HeightForWidth with MaximunWidth to know candidate height of multiline text.

Use GetHeightForWidth / GetWidthForHeight size as suggested minimun size.