[NUI] Update MeasureCallback interface for latest Dali::Toolkit
Making MeasureCallback return a struct is causing the .NET runtime on
Windows 10 messes up with argument values. We changed the signature of
MeasureCallback on Dali::Toolkit to make the return value be an out
parameter.

The detailed explanation of what's the problem can be found at [1].

Moreover, we declare MeasureCallback as Cdecl, as it is declared this
way in the C++ header file.

[1] https://review.tizen.org/gerrit/#/c/platform/core/uifw/dali-toolkit/+/242665/