[NUI] Add TabView and related classes
TabView and its related classes, TabBar, TabContent, TabButton and
TabButtonGroup, are added.

TabView consists of TabBar and TabContent.

TabBar consists of a list of TabButton.
TabContent consists of a list of View.

TabView.AddTab() adds TabButton and View to TabBar and TabContent in
TabView.
TabView.RemoveTab() removes TabButton and View from TabBar and
TabContent in TabView.