[NUI] Fix recursion when the view sets KeyboardNavigationSupport to true
In Dali code, the method GetNextKeyboardFocuableActor should return
an empty actor (= the base handle without a referenced object)
if there is no actor can be focused.

The corresponding method in NUI is GetNextFocusableView.
It returned new View() (that is not an empty actor) which leads to the wrong result.
So instead returning new View, this patch returns a null IntPtr.

Signed-off-by: Jiyun Yang <ji.yang@samsung.com>