Revert "[NUI] Fix CA2000 issue in DefaultTheme(Dispose object before losing scope)"
This reverts commit cff3fd4d9fd6fd461263e742e296a5397e9563a0.

This causes segfault to dispose styles after passing them through `AddStyleWithoutClone` method.
Because the theme object holds the passed style objects.
Which means the responsibility to dispose the object is transferred to the theme object.

In this case, we can suppress this warning.
https://docs.microsoft.com/en-us/dotnet/fundamentals/code-analysis/quality-rules/ca2000

Instead revert the patch, it suppressed CA2000 warning.