[MachineLearning.Inference][Non-ACR] Fix the potential bug (#2547)
* [MachineLearning.Inference] Update the version number of the shared library

This patch updates the version number of the nnstreamer library.
* libcapi-nnstreamer.so.0 -> libcapi-nnstreamer.so.1

Signed-off-by: Sangjung Woo <sangjung.woo@samsung.com>

* [MachineLearning.Inference] Remove the name comparision in TensorsInfo.Equals()

Since the name of the TensorsInfo is not used when negociating the
pipeline, the name comparison in TensorsInfo.Equals() should be removed.

Signed-off-by: Sangjung Woo <sangjung.woo@samsung.com>

* [MachineLearning.Inference] Add Clone() to TensorsInfo class for deep copy

It is not a common case but developers can modify the TensorsInfo after
creating the TensorsData. Because of this reason, TensorsData should
have a hard-copied instance of the TensorsInfo object. If not, it shows
the wrong information of TensorsData. This patch fixes this potential
bug.

Signed-off-by: Sangjung Woo <sangjung.woo@samsung.com>