[NUI] : Add ScrollPosition in ScrollEventArgs for user convenience.
Current Position property in ScrollEventArgs is position of
ContentContainer which is negative.
user controls scrollableBase via ScrollTo and other API with
positive position of pan not ContentContainer so user must convert
current Position Property to use.

By this patch, we provide positive ScrollPosition which represent
position of scroll pan, so no need to convert value for using it.