[NUI] TextLayout can consider specified maximum width.
Previsouly, the Layout system did not reflect TextLabel's maximum size, so the following example did't work corrrectly,

```C#
  var view = new View()
  {
    Layout = new LinearLayout() { LinearOrientation = LinearLayout.Orientation.Horizontal, },
  };
  view.Add(new TextLabel() { MaximumSize = new Size2D(50, 50), Text = "Helloooooooooooooooooo!" });
  view.Add(new TextLabel() { MaximumSize = new Size2D(50, 50), Text = "Helloooooooooooooooooo!" });
  NUIApplication.GetDefaultWindow().Add(view);
```

The x-position of the second child was calculated based on the first child's natutal width,
but the first child's actual width was limited by MaximumSize(50).

So this patch update the TextLabel's layout to make it consider the maxmimum size of the view.

Signed-off-by: Jiyun Yang <ji.yang@samsung.com>