[NUI] : Scrollbar thumb support image and update styles
Change thumb View to ImageView to support image and color properly.

Add new property for thumb images on Scrollbar and ScrollbarStyle.
- ThumbVerticalImageUrl
- ThumbHorizontalImageUrl

Add new image resource for thumb image.
- nui_component_default_scroll_vbar.#.png
- nui_component_default_scroll_hbar.#.png

Now color can be color of background for thumb when image is not
existed, or color of image if url is existed.