[NUI] Dialog and AlertDialog code refactoring with adding DialogPage
AlertDialog is no longer derived from Dialog to manage easily.
Because it is difficult that AlertDialog supports Dialog.Content with
AlertDialog's TitleContent and ActionContent.

Dialog and AlertDialog's constructors with parameters are removed.

AlertDialog's Actions is added to set these properties with constructor.

AlertDialog's SetActionButton() is removed because the above Actions
covers this method.

DialogPage class supports dialog page features as follows.
- Content
- Scrim

ShowDialog() and ShowAlertDialog() is moved from Navigator class to
DialogPage class.