[NUI] Deprecate ResizePolicyType
Up to now, both ResizePolicyType and LayoutParamPolicies are supported.
Originally, ResizePolicyType is for a view whose parent view does not
have Layout.
In contrast, LayoutParamPolicies is for a view whose parent view has
Layout.

However, ResizePolicyType can be used as LayoutParamPolicies as well
internally.

This makes users confused which one they need to use in each case.

Not to confuse users, ResizePolicyType is deprecated.

Instead, LayoutParamPolicies can be used to set the size of a view
relative to its parent or children.

Specifically, LayoutParamPolicies.MatchParent and
LayoutParamPolicies.WrapContent work with parent view which does not
have Layout as how ResizePolicyType.FillToParent and
ResizePolicyType.FitToChildren work.