[NUI] Fix to calculate RelativeLayout's grand children position correctly
Previously, RelativeLayout set its children View's size and position but the
children Layout's MeasuredWidth/Height did not match to the children View's size
and position.
This causes that the grand children's position was calculated incorrectly
because the grand children's position was calculated based on the children's
MeasuredWidth/Height.

Now, RelativeLayout sets its children Layout's MeasuredWidth/Height based on the
children View's size and the grand children's position is calculated correctly.