[NUI] Fix to support DimensionDependency in Layout
Previously, View with DimensionDependency could not be calculated
correctly by Layout.
Because the View's ResizePolicy is changed from DimensionDependency to
Fixed when Layout sets the View's Size.

Now, to support View with DimensionDependency, Layout does not change
width or height with DimensionDependency.
Instead, Layout calculates Views when a View is resized to use the size
calculated from DimensionDependency in Dali.