[NUI] Add AdjustViewSize to ImageView
Up to now, although the position and size of image resource of ImageView
is set by FittingMode, but there is no way to set the size of ImageView
to respect the aspect ratio of its image resource.

If AdjustViewSize is true, then ImageView respects the aspect ratio of
its image resource when it determines its size.

AdjustViewSize works only if the ImageView is added to a View having Layout.
Otherwise, use can set ImageView's size to respect the aspect ratio of its image
resource by using ResizePolicyType.DimensionDependency.