[NUI] Fix ContentPage to calculate AppBar and Content size correctly
Previously, the size of AppBar and Content was set by setting Size2D
property.
This caused AppBar and Content to change their
Width/HeightSpecification.
Consequently, the size of AppBar and Content was not recalculated when
window size was changed.

Now, the size of AppBar and Content is set by SetSize() method.
Since SetSize() method does not change Width/HeightSpecification, the
size of AppBar and Content is recalculated correctly when window size is
changed.