Support system default CultureInfo
When the system locale is "tl_PH.utf8", the CoreApplication sets the
CultureInfo that is created using "en". Because, there is no CultureInfo
about "tl-PH" and "tl". But, the developer can create the resource using
"tl-PH". In this case, developers expect the name of the CultureInfo is "tl-PH".
Platform should set the current CultureInfo that is "tl-PH".
This patch supports system default CultureInfo using xml file.
Platform developers can set the default culture name for each locale.
The CoreApplication tries to create the CultureInfo using the culture name.

Signed-off-by: Hwankyu Jhun <h.jhun@samsung.com>