[NUI] Ignore ResizePolicy when View size is calculated by Layout
View size is calculated based on Width/HeightSpecification if View's
Layout is set.

However, if View's ResizePolicy is not UseNaturalSize (default value),
then View's ResizePolicy also affects View size calculation in DALi
although View's Layout is set.

To ignore ResizePolicy when View size is calculated by View's Layout,
ResizePolicy is set with FIXED when View's Layout is set.

ResizePolicy is stored when Layout is set and it is ignored when View
size is calculated.
ResizePolicy is restored when Layout is unset and it is considered when
View size is calculated.