[NUI] Add ContentPageLayout to calculate Content size correctly
Previously, ContentPage's Layout was AbsoluteLayout so the height of
Content was calculated to the same with the height of ContentPage,
although AppBar existed at the top of ContentPage.

Now, ContentPageLayout has been introduced and ContentPageLayout
calculates the height of Content except the height of AppBar.
Consequently, OnRelayout() is no longer required because
ContentPageLayout calculates the size and position of Content.