[NUI] Add TextFieldLayout and TextEditorLayout
In DALi, the default HeightResizePolicy of TextField and TextEditor is
FillToParent.
Because of this, TextField and TextEditor fill their parent's height by
default although text is null string.

If TextField and TextEditor's parent has Layout, then their sizes should
be calculated based on their Width/HeightSpecification.

To calculate TextField and TextEditor's size based on their
Width/HeightSpecification, TextFieldLayout and TextEditor are added.