[NUI][AT-SPI] Add AccessibilityStatesNotifyMode
The different semantics of the second (bool) parameter of
EmitAccessibilityStatesChangedEvent and NotifyAccessibilityStatesChange
can cause confusion, since it represents the new value of the state in
the former method, and the notification strategy in the latter.

This change introduces an enum type for the second parameter of
NotifyAccessibilityStatesChange to express its effect more clearly.