[NUI] Get native image PropertyMap instead of cached.
When we try to get cached property map for animated image,
It will not return internal-updated values
something like "CurrentFrame" for animated image.

This patch make we return Current PropertyMap always
instead of cached property map.

Actually, the best way to fix is, create some flags when we
try to get time-sensitive values.
But it need to fixup some codes and test. So just make hot-fix.

Signed-off-by: Eunki, Hong <eunkiki.hong@samsung.com>