[NUI][AT-SPI] Support multiple AT-SPI interfaces for View
This patch adds GetInterfaces to AccessibilityDelegate which will allow
to report any number of AT-SPI interfaces in any combinations back to
DALi. Complementary changes to DALi make the AT-SPI bridge use
GetInterfaces instead of dynamic_cast (which stops at the C++ language
boundary).

Note however, that it is not yet possible to declare more than one
AT-SPI interface for a given Control in NUI. There are more changes
needed, such as removing SetAccessibilityConstructor and switching to
native C# interfaces, and they will be done in separate PRs.