[DO NOT REVIEW] [NUI] Add a HitTestResult event.
In the case of TouchEvent or Gesture, there is no way to propagate the event to the view below that is not related.

So, before sending an touch event, send an HitTestResult event to the view in the hit-test process to ask whether it will be hit or not.
If it returns false, it means that it will not be hit, and the hit-test continues to the next view.
If it returns true, it means that it was hit, and the event is called from the view as before, and the event is propagated to the parent.