[NUI][AT-SPI] Remove interops for SetAccessibilityConstructor()
The method name SetAccessibilityConstructor() in NUI has always been
misleading, because it actually never had an 'accessibilityConstructor'
parameter in the first place. With the recent addition of bindings for
Accessible::GetInterfaces(), all NUI controls use the same Accessible
implementation (NUIViewAccessible) anyway. However, as this method is
currently used in many places, it is kept as a wrapper for setting
AccessibilityRole (and interface flags) and will be removed if/when we
decide to switch to AT-SPI interfaces as native C# interfaces. Then, all
usages of SetAccessibilityConstructor() would be replaced by simply
setting AccessibilityRole (and interface flags won't be needed).