[NUI] Save the cropMask value internally in ImageView
Regardless of the cropToMask setting order, the value has been modified to apply properly.
In the existing code, the appclication  set the cropToMask first and setting url later,
there was an issue that the set cropToMask value was not applied properly.
To solve this issue, it has been modified to save this value at the ImageView.