[NUI] Supports Clockwise and CounterClockwise focus movement.
Clockwise and CounterClockwise are rotary wheel event directions.

There is also a requirement that the focus be moved with a rotary wheel as well.

So, you can set the FocusableView according to the clockwise and counterclockwise directions of the rotary wheel.

```C#
  var viewA = new View { Focusable = true };
  var ViewB = new View { Focusable = true };

  viewA.ClockwiseFocusableView = viewB;
  ViewB.CounterClockwiseFocusableView = ViewA;

```

In ViewA, if the rotary wheel is in the Clockwise direction, ViewB is focused.
When the rotary wheel in ViewB is in the CountClockwise direction, ViewA is focused.

Dependency
https://review.tizen.org/gerrit/#/c/platform/core/uifw/dali-toolkit/+/272040/
https://review.tizen.org/gerrit/#/c/platform/core/uifw/dali-csharp-binder/+/272137/