[NUI] Fix MarginProperty and PaddingProperty with Layout
View.Margin and Padding refers View.Layout.Margin and Padding.

So if View.Margin and Padding are set with Layout, then
View.Layout.Margin and Padding are set and View.Margin and Padding
become Extents(0, 0, 0, 0).

Previously, the problem was that View.MarginProperty and PaddingProperty
did not support the above logic properly.

Now, to resolve the above issue, View.MarginProperty and PaddingProperty
support the above logic properly.
i.e. Setting View.MarginProperty and PaddingProperty with Layout set
     value to View.Layout.Margin and Padding and also set
     Extents(0, 0, 0, 0) to View.Margin and Padding.