[NUI] Make the velocity of slider as 1%
Previous code can move only 1 units when we try to change value
by Key event.

This patch make the value moved for 1%. It will change as same lates
even min/max value difference is big, or small.

TODO : Currently, we make it as const value 1%.
Should we need to make this value as property?

Signed-off-by: Eunki, Hong <eunkiki.hong@samsung.com>