[NUI] Propagate AutomationId value to DALi (#4426)
Co-authored-by: Artur Świgoń <a.swigon@samsung.com>

Due to an oversight in #4389, the automationId attribute disappeared from the AT-SPI tree for some Views, causing multiple regressions in Aurum tests. This PR fixes this bug.

The "automationId" key set in AccessibilityAttributes is ignored for controls not backed by a NUIViewAccessible (e.g. BaseComponents, FluxView etc.), because NUIViewAccessible::GetAttributes is never called for such controls. The value of AutomationId should be propagated to dali-toolkit instead, so that is it visible to ControlAccessible::GetAttributes, therefore having an effect for all types of Views (NUIViewAccessible::GetAttributes calls ControlAccessible::GetAttributes first, and then collects any additional attributes stored in View.AccessibilityAttributes).

Dependencies:

    https://review.tizen.org/gerrit/#/c/platform/core/uifw/dali-toolkit/+/278352/ (merged)
    https://review.tizen.org/gerrit/#/c/platform/core/uifw/dali-csharp-binder/+/278353/ (merged)