[NUI.XamlBuild] Sync Xaml parser code in Tizen.NUI to Tizen.NUI.Xaml
Now we want to support building Xaml for themes at application compile time,
it is needed to sync some code lines previously added to Tizen.NUI's xaml parser to Tizen.NUI.Xaml.

The contents of the added code are as follows:
When a child has a string key and a parent node has a proper type of indexer,
it trys to treat a parent node as a dictionary type which has items of children nodes.

Signed-off-by: Jiyun Yang <ji.yang@samsung.com>