[NUI] Refactor Signal structure.
Previously, When NUI wanted to connect or disconnect a handler to the Signal,
it had to create a Signal instance just for it and dispose the instance right after.
Repeating this over and over again isn't cheap.
And also NUI had to declare many Signal classes with almost idential structures.

This patch provides Signal connection/disconnection pinvoke functions
to make it possible to remove redundant signal classes and instance creations.

Signed-off-by: Jiyun Yang <ji.yang@samsung.com>