[NUI][AT-SPI] Override ContentPage.AccessibilityGetName()
ContentPage often serves as a default label in scenarios where Navigator is
used, so Screen Reader will ask ContentPage about the accessibility name (text
to read) when a Navigator page is pushed. Therefore, it makes sense to try to
provide the text from ContentPage.AppBar.Title (if it exists). The application
developer can always customize this by setting the AccessibilityName property
or adding an AccessibilityRelation.LabelledBy.