[NUI][AT-SPI] Add ViewAccessibilityMode (#5118)
The accessibility mode helps select the right Accessible implementation for
a given view. It does not affect controls derived from CustomView (or any of
its subclasses), e.g. Tizen.NUI.Components, which already behave in the custom
mode. The default mode is kept for Tizen.NUI.BaseComponents. However, this
patch add a possibility of opting into the custom mode for controls derived
directly from View (e.g. FluxView), and by doing so, to be able to implement
accessibility support for them in the C# language.

Co-authored-by: Artur Świgoń <a.swigon@samsung.com>