[NUI] Make subclass of ImageView can use ResourceUrl normally.
If we skip to set ImageVisualType, or set ImageVisualType as Image,
we will try to find best visual by it's file format codes.
But previously, we fix the type of visual.
So, instead of some special case like URLs, just make we use Image type.

And also, make _resourceUrl latest value if we success to get url from visual.

Signed-off-by: Eunki Hong <eunkiki.hong@samsung.com>