[NUI][API10] Cherry-pick Registry relative codes
This is a combination of 6 commits.

- This is the 1st commit message:

[NUI] Let registry throw exception at error case

Let we throw exception if same object try to register.

Moreover, let we print more useful callstack if we got unmatched case.

Signed-off-by: Eunki, Hong <eunkiki.hong@samsung.com>

- This is the commit message #2:

[NUI] Consider WeakReference resurrection at Registry.

Since NUI use DisposeQueue system instead of GC directly,
the lifecycle of WeakReference may need to be alive more times.

If we make trackResurrection as false, WeakReference.Target
become null when the item is in DisposeQueue.

Signed-off-by: Eunki Hong <eunkiki.hong@samsung.com>

- This is the commit message #3:

[NUI] Temperary revert exception throw at Registry.

Until resolve all kind of apps fix wrong behavior,
let we just print error without exception.

Signed-off-by: Eunki, Hong <eunkiki.hong@samsung.com>

- This is the commit message #4:

[NUI] Do not register when BaseHandle doesn't have body

If BaseHandle doesn't have body, the refCPtr value is IntPtr.Zero.

If than, the various cases have wrong registration in future.

To avoid this case, let we ignore to register to registry, and
make BaseHandle.registerMe value as false.

Signed-off-by: Eunki, Hong <eunkiki.hong@samsung.com>

- This is the commit message #5:

[NUI] Resolve the potential issue when DisposeQueued handle get

It is possible that DisposeQueued handle be used as input of GetManagedBaseHandleFromRefObject.

In this case, even if app hold that BaseHandle, it will be Disposed by DisposeQueue.
If then the returned item become disposed, and it might breakdown app logics.

Signed-off-by: Eunki, Hong <eunkiki.hong@samsung.com>

- This is the commit message #6:

[NUI] Unregister disposed / disposequeued BaseHandle and register again if need

It is possible that dispose queued basehandle try to get / register again
due to some reason. Like, Native side object is alived but C# side has no reference.

We already to to detect that cases. But if BaseHandle added at DisposeQueue,
we cannot avoid to call the Dispose(); after some event loops later, and then
Unregister the handle.

To avoid some error cases, let we Unregister the basehandle synchronously
and then re-register base on the new BaseHandle.

Signed-off-by: Eunki, Hong <eunkiki.hong@samsung.com>